N21-2175 Add course boards to bettermarks (#5238)
