+++ apps/server/src/modules/board/service/media-board/media-board.service.ts
@@ -1,16 +1,15 @@
-import { Injectable } from '@nestjs/common';
-import { EntityId } from '@shared/domain/types';
+import { ObjectId } from '@mikro-orm/mongodb';
 
 import { ToolContextType } from '@modules/tool/common/enum';
-import { ContextExternalToolService } from '@modules/tool/context-external-tool/service';
 import { ContextExternalTool, ContextRef } from '@modules/tool/context-external-tool/domain';
+import { ContextExternalToolService } from '@modules/tool/context-external-tool/service';
 import { SchoolExternalTool, SchoolExternalToolRef } from '@modules/tool/school-external-tool/domain';
-import { ObjectId } from '@mikro-orm/mongodb';
+import { Injectable } from '@nestjs/common';
+import { EntityId } from '@shared/domain/types';
 import {
 	AnyMediaBoardNode,
 	BoardExternalReference,
 	isMediaBoard,
-	isMediaExternalToolElement,
 	MediaBoard,
 	MediaExternalToolElement,
 } from '../../domain';
@@ -57,27 +56,14 @@ export class MediaBoardService {
 			schoolToolRef: { schoolToolId: schoolExternalTool.id },
 		});
 
-		const existing = this.findMediaElements(mediaBoard);
+		const existingExternalToolElements: MediaExternalToolElement[] =
+			mediaBoard.getChildrenOfType(MediaExternalToolElement);
 
-		const exists = existing.some((element) =>
-			contextExternalTools.some((tool) => tool.id === element.contextExternalToolId)
+		const exists: boolean = existingExternalToolElements.some((element: MediaExternalToolElement): boolean =>
+			contextExternalTools.some((tool: ContextExternalTool): boolean => tool.id === element.contextExternalToolId)
 		);
 
 		return exists;
-	}
-
-	public findMediaElements(boardNode: AnyMediaBoardNode): MediaExternalToolElement[] {
-		const elements = boardNode.children.reduce((result: MediaExternalToolElement[], bn) => {
-			result.push(...this.findMediaElements(bn as AnyMediaBoardNode));
-
-			if (isMediaExternalToolElement(bn)) {
-				result.push(bn);
-			}
-
-			return result;
-		}, []);
-
-		return elements;
 	}
 
 	public async updateBackgroundColor<T extends WithBackgroundColor<AnyMediaBoardNode>>(
