+++ apps/server/src/modules/pseudonym/service/index.ts
@@ -1,2 +1,1 @@
 export * from './pseudonym.service';
-export * from './feathers-roster.service';
