+++ apps/server/src/modules/roster/index.ts
@@ -1,0 +1,1 @@
+export { FeathersRosterService } from './service';
