+++ apps/server/src/modules/roster/roster.config.ts
@@ -1,0 +1,3 @@
+export interface RosterConfig {
+	FEATURE_COLUMN_BOARD_EXTERNAL_TOOLS_ENABLED: boolean;
+}
