+++ apps/server/src/modules/roster/service/feathers-roster.service.ts
@@ -1,6 +1,7 @@
 import { CourseService } from '@modules/learnroom/service';
+import { PseudonymService } from '@modules/pseudonym/service';
 import { ToolContextType } from '@modules/tool/common/enum';
-import { ContextExternalTool, ContextRef } from '@modules/tool/context-external-tool/domain';
+import { ContextExternalTool } from '@modules/tool/context-external-tool/domain';
 import { ContextExternalToolService } from '@modules/tool/context-external-tool/service';
 import { ExternalTool } from '@modules/tool/external-tool/domain';
 import { ExternalToolService } from '@modules/tool/external-tool/service';
@@ -8,14 +9,17 @@ import { SchoolExternalTool } from '@modules/tool/school-external-tool/domain';
 import { SchoolExternalToolService } from '@modules/tool/school-external-tool/service';
 import { UserService } from '@modules/user';
 import { Injectable } from '@nestjs/common';
+import { ConfigService } from '@nestjs/config';
 import { NotFoundLoggableException } from '@shared/common/loggable-exception';
 import { Pseudonym, RoleReference, UserDO } from '@shared/domain/domainobject';
 import { Course } from '@shared/domain/entity';
 import { RoleName } from '@shared/domain/interface';
 import { EntityId } from '@shared/domain/types';
-import { PseudonymService } from './pseudonym.service';
+import { BoardExternalReferenceType, ColumnBoard, ColumnBoardService } from '../../board';
+import { ExternalToolElement } from '../../board/domain';
+import { RosterConfig } from '../roster.config';
 
-interface UserMetdata {
+interface UserMetadata {
 	data: {
 		user_id: string;
 		username: string;
@@ -60,14 +64,16 @@ export class FeathersRosterService {
 		private readonly courseService: CourseService,
 		private readonly externalToolService: ExternalToolService,
 		private readonly schoolExternalToolService: SchoolExternalToolService,
-		private readonly contextExternalToolService: ContextExternalToolService
+		private readonly contextExternalToolService: ContextExternalToolService,
+		private readonly columnBoardService: ColumnBoardService,
+		private readonly configService: ConfigService<RosterConfig, true>
 	) {}
 
-	async getUsersMetadata(pseudonym: string): Promise<UserMetdata> {
+	public async getUsersMetadata(pseudonym: string): Promise<UserMetadata> {
 		const loadedPseudonym: Pseudonym = await this.findPseudonymByPseudonym(pseudonym);
 		const user: UserDO = await this.userService.findById(loadedPseudonym.userId);
 
-		const userMetadata: UserMetdata = {
+		const userMetadata: UserMetadata = {
 			data: {
 				user_id: user.id as string,
 				username: this.pseudonymService.getIframeSubject(loadedPseudonym.pseudonym),
@@ -78,12 +84,8 @@ export class FeathersRosterService {
 		return userMetadata;
 	}
 
-	async getUserGroups(pseudonym: string, oauth2ClientId: string): Promise<UserGroups> {
-		const loadedPseudonym: Pseudonym = await this.findPseudonymByPseudonym(pseudonym);
-		const externalTool: ExternalTool = await this.validateAndGetExternalTool(oauth2ClientId);
-
-		let courses: Course[] = await this.getCoursesFromUsersPseudonym(loadedPseudonym);
-		courses = await this.filterCoursesByToolAvailability(courses, externalTool.id);
+	public async getUserGroups(pseudonym: string, oauth2ClientId: string): Promise<UserGroups> {
+		const courses: Course[] = await this.getCourses(pseudonym, oauth2ClientId);
 
 		const userGroups: UserGroups = {
 			data: {
@@ -100,12 +102,31 @@ export class FeathersRosterService {
 		return userGroups;
 	}
 
-	async getGroup(courseId: EntityId, oauth2ClientId: string): Promise<Group> {
-		const course: Course = await this.courseService.findById(courseId);
-		const externalTool: ExternalTool = await this.validateAndGetExternalTool(oauth2ClientId);
+	private async getCourses(pseudonym: string, oauth2ClientId: string): Promise<Course[]> {
+		const pseudonymContext: Pseudonym = await this.findPseudonymByPseudonym(pseudonym);
+		const user: UserDO = await this.userService.findById(pseudonymContext.userId);
 
-		await this.validateSchoolExternalTool(course.school.id, externalTool.id);
-		await this.validateContextExternalTools(courseId);
+		const externalTool: ExternalTool = await this.validateAndGetExternalTool(oauth2ClientId);
+		const schoolExternalTool: SchoolExternalTool = await this.validateSchoolExternalTool(
+			user.schoolId,
+			externalTool.id
+		);
+
+		let courses: Course[] = await this.courseService.findAllByUserId(pseudonymContext.userId);
+		courses = await this.filterCoursesByToolAvailability(courses, schoolExternalTool);
+
+		return courses;
+	}
+
+	public async getGroup(courseId: EntityId, oauth2ClientId: string): Promise<Group> {
+		const course: Course = await this.courseService.findById(courseId);
+
+		const externalTool: ExternalTool = await this.validateAndGetExternalTool(oauth2ClientId);
+		const schoolExternalTool: SchoolExternalTool = await this.validateSchoolExternalTool(
+			course.school.id,
+			externalTool.id
+		);
+		await this.validateContextExternalTools(course, schoolExternalTool);
 
 		const [studentEntities, teacherEntities, substitutionTeacherEntities] = await Promise.all([
 			course.students.loadItems(),
@@ -163,35 +184,21 @@ export class FeathersRosterService {
 		return loadedPseudonym;
 	}
 
-	private async getCoursesFromUsersPseudonym(pseudonym: Pseudonym): Promise<Course[]> {
-		const courses: Course[] = await this.courseService.findAllByUserId(pseudonym.userId);
-
-		return courses;
-	}
-
-	private async filterCoursesByToolAvailability(courses: Course[], externalToolId: string): Promise<Course[]> {
+	private async filterCoursesByToolAvailability(
+		courses: Course[],
+		schoolExternalTool: SchoolExternalTool
+	): Promise<Course[]> {
 		const validCourses: Course[] = [];
 
 		await Promise.all(
-			courses.map(async (course: Course) => {
-				const contextExternalTools: ContextExternalTool[] = await this.contextExternalToolService.findAllByContext(
-					new ContextRef({
-						id: course.id,
-						type: ToolContextType.COURSE,
-					})
+			courses.map(async (course: Course): Promise<void> => {
+				const isExternalToolReferencedInCourse: boolean = await this.isExternalToolReferencedInCourse(
+					course,
+					schoolExternalTool
 				);
 
-				for await (const contextExternalTool of contextExternalTools) {
-					const schoolExternalTool: SchoolExternalTool = await this.schoolExternalToolService.findById(
-						contextExternalTool.schoolToolRef.schoolToolId
-					);
-					const externalTool: ExternalTool = await this.externalToolService.findById(schoolExternalTool.toolId);
-					const isRequiredTool: boolean = externalTool.id === externalToolId;
-
-					if (isRequiredTool) {
-						validCourses.push(course);
-						break;
-					}
+				if (isExternalToolReferencedInCourse) {
+					validCourses.push(course);
 				}
 			})
 		);
@@ -199,36 +206,107 @@ export class FeathersRosterService {
 		return validCourses;
 	}
 
+	private async isExternalToolReferencedInCourse(
+		course: Course,
+		schoolExternalTool: SchoolExternalTool
+	): Promise<boolean> {
+		const contextExternalToolsInCourse: ContextExternalTool[] =
+			await this.contextExternalToolService.findContextExternalTools({
+				context: {
+					id: course.id,
+					type: ToolContextType.COURSE,
+				},
+				schoolToolRef: {
+					schoolToolId: schoolExternalTool.id,
+				},
+			});
+
+		if (contextExternalToolsInCourse.length > 0) {
+			return true;
+		}
+
+		if (this.configService.get<boolean>('FEATURE_COLUMN_BOARD_EXTERNAL_TOOLS_ENABLED')) {
+			const columnBoards: ColumnBoard[] = await this.columnBoardService.findByExternalReference({
+				type: BoardExternalReferenceType.Course,
+				id: course.id,
+			});
+
+			const isExternalToolReferencedInColumnBoard: boolean[] = await Promise.all(
+				columnBoards.map(
+					async (columnBoard: ColumnBoard): Promise<boolean> =>
+						this.isExternalToolReferencedInColumnBoard(columnBoard, schoolExternalTool)
+				)
+			);
+
+			if (isExternalToolReferencedInColumnBoard.some(Boolean)) {
+				return true;
+			}
+		}
+
+		return false;
+	}
+
+	private async isExternalToolReferencedInColumnBoard(
+		columnBoard: ColumnBoard,
+		schoolExternalTool: SchoolExternalTool
+	): Promise<boolean> {
+		const elements: ExternalToolElement[] = columnBoard.getChildrenOfType(ExternalToolElement);
+
+		const hasRequestedTool: boolean[] = await Promise.all(
+			elements.map(async (element: ExternalToolElement): Promise<boolean> => {
+				if (!element.contextExternalToolId) {
+					return false;
+				}
+
+				const contextExternalTool: ContextExternalTool | null = await this.contextExternalToolService.findById(
+					element.contextExternalToolId
+				);
+
+				const isRequestedTool: boolean = contextExternalTool?.schoolToolRef.schoolToolId === schoolExternalTool.id;
+
+				return isRequestedTool;
+			})
+		);
+
+		const hasTool: boolean = hasRequestedTool.some(Boolean);
+
+		return hasTool;
+	}
+
 	private async validateAndGetExternalTool(oauth2ClientId: string): Promise<ExternalTool> {
 		const externalTool: ExternalTool | null = await this.externalToolService.findExternalToolByOAuth2ConfigClientId(
 			oauth2ClientId
 		);
 
-		if (!externalTool || !externalTool.id) {
+		if (!externalTool || !externalTool.id || externalTool.isDeactivated) {
 			throw new NotFoundLoggableException(ExternalTool.name, { 'config.clientId': oauth2ClientId });
 		}
 
 		return externalTool;
 	}
 
-	private async validateSchoolExternalTool(schoolId: EntityId, toolId: string): Promise<void> {
+	private async validateSchoolExternalTool(schoolId: EntityId, toolId: string): Promise<SchoolExternalTool> {
 		const schoolExternalTools: SchoolExternalTool[] = await this.schoolExternalToolService.findSchoolExternalTools({
 			schoolId,
 			toolId,
+			isDeactivated: false,
 		});
 
 		if (schoolExternalTools.length === 0) {
 			throw new NotFoundLoggableException(SchoolExternalTool.name, { toolId });
 		}
+
+		return schoolExternalTools[0];
 	}
 
-	private async validateContextExternalTools(courseId: EntityId): Promise<void> {
-		const contextExternalTools: ContextExternalTool[] = await this.contextExternalToolService.findAllByContext(
-			new ContextRef({ id: courseId, type: ToolContextType.COURSE })
+	private async validateContextExternalTools(course: Course, schoolExternalTool: SchoolExternalTool): Promise<void> {
+		const isExternalToolReferencedInCourse: boolean = await this.isExternalToolReferencedInCourse(
+			course,
+			schoolExternalTool
 		);
 
-		if (contextExternalTools.length === 0) {
-			throw new NotFoundLoggableException(ContextExternalTool.name, { 'contextRef.id': courseId });
+		if (!isExternalToolReferencedInCourse) {
+			throw new NotFoundLoggableException(ContextExternalTool.name, { contextId: course.id });
 		}
 	}
