+++ apps/server/src/modules/roster/service/index.ts
@@ -1,0 +1,1 @@
+export { FeathersRosterService } from './feathers-roster.service';
