+++ apps/server/src/shared/domain/types/constructor.ts
@@ -1,0 +1,1 @@
+export type Constructor<T> = new (...args: never[]) => T;
